requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.5
langchain-xai>=0.2.1
python-dotenv>=1.0.0
